challenge: Delay bot entry into elevator shaft by gold SAS doors
When a bot is waiting by the gold SAS doors for the elevator to arrive at the upper floor, the arrival event is delayed for a second to give anyone on the elevator a moment to gracefully exit the elevator shaft before the waiting bot tries to enter it. This should help reduce bots impeding each other when trying to use this elevator.