gantry: Fix issue with bots not walking properly on the catwalk below the window
Apparently the AI "cliff check" was causing the bot to think it was unsafe to walk on the catwalk, forcing the bot to jump (via obstruction check) to navigate the catwalk. The 0x80 special info flag has been added to several sectors to disable the cliff check, which allows bots to walk properly across the entire catwalk without needing to jump.

A couple of node sizes and positions at the north end of the catwalk have also been adjusted to compensate for the lack of cliff checking. This helps keep the bots from turning too soon and trying to run along the north catwalk on the outside of the rails, causing them to fall into the pit. Previously, the cliff check helped protect against this.