fortress: Initial adaptation of "Imperial Fortress" for Rbots
Both the KFY and normal deathmatch variants of the map are supported.

In the non-KFY variant of the map, several ghost things were added to a layer named 'padding' to keep the node thing numbers consistent with the KFY variant. This simplifies copy/pasting the nodes and cogs between the two files. These padding things have been given the 0x18000 thing flags to remove them from the map and free up those thing slots when the game loads.