Refactor navigation code to use messages instead of triggers
This refactoring significantly improves performance when there are many bots and many paths. Consider, for instance, the Imperial Fortress map, which currently has 30 paths and over 120 path nodes. The gameplay started degrading with only 6 bots in the match because of all the triggers being sent to/from rbot-nav.cog. After refactoring to send direct messages instead of triggers, gameplay for the same map is smooth even with 16 bots.

Code quality/readability is also improved slightly. The code is a little more coherent now that retrieval of the response data is generally handled immediately following the request, rather than in some separate trigger handler code elsewhere in the file.

The rbot-item-special.cog was also adjusted slightly to match the updated style of rbot-nav-special.cog for consistency.