challenge: Improve so bots completely ignore rail detonator above red base entrance
This was originally implemented with special info on some of the sectors by the elevator shaft (the rbot-item-special.cog hadn't been implemented yet). However, some sectors were not flagged, so when exiting the elevator shaft, the bot would start and stop seeking items in such a way that he would go forward, back, then forward again after getting off the elevator. The special info for these sectors has been replaced with item special on the rail detonator powerup itself, so now the bot will only ignore that powerup and not others that may be in the hallway. This improves the bot behavior so he simply runs forward without turning back.