Initialize bot difficulty from player's selected difficulty
This enables the player to choose the desired bot difficulty through the game's standard difficulty setting checkboxes at player creation time. This is especially helpful to control the difficulty when the Rbot Admin is not available (e.g. playing with mods).

The difficulty symbol in rbot-startup.cog has been made local and can no longer be initialized by passing a value from the JKL.