Cleanup: Remove unused/unnecessary code from rbot-srv-weapons.cog
Note: The ParseArg/eyeoffset code is being removed because it does not have the intended effect. When applying the eyeoffset, the game considers only the bot's position and eyePYR (head tilt), but not the bot's LVec (the direction the bot is facing). In other words, the eye gets offset as if the bot were facing forward (0/1/0). The intention was for LOS to be computed from the barrel of the current gun, but since the LVec is not included in the eyeoffset calculations, adding any X/Y axis offset does not yield the expected result.