Cleanup: Consolidate saber projectile timers
The timers were identical except for the projectile fired. They have been consolidated into one timer that accepts the projectile as a parameter. This also makes the code a little more readable since the projectiles fired are now specified in the code that uses them, rather than being separated out in a separate block of timer code.