Add weapon/ammo arrays in backpack cogs
Instead of using one large 'item0' array to hold the data for all backpack contents, the backpack cogs now use a separate array for each item that can be contained within the backpack. This makes the code more readable, as the symbol names are now self-describing and no longer rely on index+offset calculations to find the correct data slot.