Clean up Rbot Admin stats output
The Rbpt Admin stats output has been reordered to improve organization and
readability. Related items are now grouped together a little better.

Additionally, the stats display will be redrawn almost immediately after a
bot or player dies. This minimizes visual disruption of the stats display
when the death/score text appears on the screen and scrolls away the stats
text. This makes it a little more pleasant to continuously monitor a bot's
stats without having the text jump around on the screen if bots are scoring
frequently. The visual disruption of the stats now only occurs within the
frame that the death/score text is first displayed on the screen, so it
appears only as a momentary "flicker" of the text.