Implement limited support for original LEC force_*.cog scripts
THe rbot-srv-force.cog scripts now support interaction with the original LEC
force_*.cog scripts. The patched scripts should be used when possible for the
"best" or "most authentic" experience, but in cases where the patched scripts
cannot be used (for example, when mods are active), the rbot-srv-force.cog
implementation will now approximate the proper behavior to offer an
"acceptable" experience. This is a best-effort approximation, but due to
limitations in how the original Force Power scripts are implemented, there are
some imperfections, such as being able to Force Blind a bot that currently has
Force Seeing activated. Using the patched Force Power scripts eliminates these
imperfections.