Fix bot lightsaber left/right slash attacks
The logic for determining whether the target is to the right or left of
the bot has been corrected so it does not use relative vector math on
an absolute vector. Also, the view dot check for determining whether
the target is in front of the bot has been adjusted to an appropriate
sensitivity. Bots are always trying to face their target, so the dot
will almost always indicate that the target is in front of the bot
unless a very sensitive threshold is used.