Allow game engine to handle blocking projectiles with the lightsaber
The 0x2000 actor flag is what the game engine uses to make an actor block saber
attacks and weapon projectiles with the 0x400 weapon flag. This feature is used
where possible to have the game engine perform blocking with the lightsaber
instead of reimplementing this behavior in the rbot-srv-weapons.cog.

In MotS, actors are unable to block projectiles with the lightsaber. Only
blocking saber attacks works correctly. This is likely a bug that was
introduced by LEC during development of MotS. To work around this, the MotS
variant of rbot-srv-weapons.cog still contains some code to emulate the saber
blocking behavior that the game engine would normally provide. This code has
also been improved from the previous implementation to be more faithful to the
behavior in JK. For example, ricochet sounds are now played when the projectile
is deflected, and the same projectile is preserved (rather than creating an
entirely new projectile) so its explosion is not created when the bot is hit.