Update SDK contents to better match original release/tutorial
The SDK archives now contain all vanilla Rbots sources from the corresponding
devel directory with the episode-specific assets omitted (JKLs, custom cogs,
etc.). This is more in line with how the SDK and tutorial from the original
release of Rbots were authored. In a previous commit, I had omitted assets from
the SDK that were included in the patch GOB/GOO containers to reduce
duplication, but this makes it more difficult to use the SDK to update external
projects that are set up the same way as Rbots itself with all the assets
unpacked from the containers.