Clean up bot saber attacks and create sparks when blade collides
Bot saber attacks now all use a single template. The 0x2 projectile flag is
used with FireProjectile to scale the damage of the projectile so the attack
will inflict the appropriate amount of damage. The damage for each bot saber
attack has been adjusted to match the player saber attacks.

The timing used to fire the projectile has also been adjusted so each attack is
fired in sync with the bot's saber blade hitting the target.

For the bot's left/right strafe saber attacks, the projectile is now fired at a
+/- 40 degree angle so it will more reliably collide with whatever is beside
the bot.

Bot saber attacks now create the appropriate sparks when the blade collides
with something. The attack will create wall sparks for surface or non-creature
thing collision, flesh sparks when damaging a bot or player, or saber block
sparks when the bot or player successfully blocks the attack with their own
saber. This improves both the visual effects and sounds of bot saber blade
collisions to more closely mimic those of an ordinary player.

Also removed some unused/obsolete templates for bot attacks/effects.

Known issue:

When bot hits a dead player/bot with a saber attack, it will create wall sparks
instead of flesh sparks. This is because the conversion from wall sparks to
flesh/saber sparks is done in the damaged: handler for the player/bot, but the
game does not run that message when a thing has the 0x200 (dead) thing flag. At
the moment, I do not have a good workaround for this.