Let game engine deflect projectiles in MotS
In JK, simply having the 0x2000 actor flag is sufficient to block deflectable
projectiles (i.e. projectiles with 0x400 weapon flag). In MotS, an actor must
additionally handle the preblock message in its class cog or in a linked cog
placed in the level and return a non-zero value, else the projectile will not
be blocked.

The workaround to recreate the game's projectile deflection behavior has been
removed from rbot-srv-weapons.cog. Now, rbot-srv.cog handles the preblock
message, which allows the bots to deflect projectiles in MotS using the
built-in functionality of the game engine.

Note: This commit introduces JK/MotS specific variants of rbot-srv.cog. This
cog is no longer identical in the JK and MotS source trees.

Also cleaned up the JK variants to match the applicable changes to MotS and
also to add a missing timer handler that plays blocking animations when a bot
saber attack is blocked.