Add support for path weight adjustment
Path weight can now be adjusted via the 'weight' parameter for rbot-path.cog.
The weight is used when a bot reaches a node that intersects with other paths.
By default, all paths have equal weight, so each path has the same (1 in N)
chance of being used. Adjusting the weight will increase or decrease the chance
of the path being selected. Paths with more weight are more likely to be
selected than paths with less weight.