spaceport: Adjust weight for spur paths near the raildet room area
With default weighting, the bots have a tendency to keep going into the same
rooms down by the raildet room area, rather than coming back up to the main
path to explore the rest of the map. This is partially due to the density of
spur paths in that area. The outdoor path is also a dead end there, so the bots
turn around and make the same random decision, again, for whether to go off
into the side rooms. These path weights have been reduced to help encourage the
bot to spend a little more time on the outdoor path and not spend an excessive
amount of time repeatedly navigating the same paths near the dead end.