Fix dodge strafe vector calculation
The old calculation incorrectly used an absolute direction in a relative
context. As a result, the bot would "strafe" relative to its map position
rather than relative to the direction it is facing. The bot's RVec, which is
relative to the LVec, is now used to implement the dodge strafing.