Reduce jump code duplication
The the bot reaches a node, the srv cog sends out a trigger to request special
info. If a special info response is received that instructs the bot to jump,
the handling code now simply sets the jump flag and lets the main AI pulse move
code handle the jumping. Previously, the handler needed to duplicate the jump
code so the jump would be handled immediately, rather than in the following AI
pulse. More recent refactoring has changed the execution order of the AI pulse
code so the special info is now handled before the move code executes, so this
duplication is no longer necessary.