arena: Add some path nodes to mitigate lost bots falling into pits
There are health and shield powerups on the lower level that did not have LOS
with any path nodes, despite being right next to the path where the bots would
be travelling. After collecting these powerups, the bots would become 'lost'
because they no longer had LOS with the node they were headed for. This could
cause the bots to needlessly wander and jump into the pit. A couple of path
nodes have been added near these powerups so the bots may return to the path
and resume navigation instead of becoming 'lost' and wandering randomly.