canopy: Initial adaptation of "Tree Canopy" level for Rbots
Several path nodes have been marked as mine placement locations. Bots will
place proximity mode sequencer charges upon reaching these path nodes.

Bots are able to obtain the concussion rifle powerup that is beneath the
elevator. This is accomplished by omitting the typical use of rbot-arrival.cog
that would normally force the bot to wait for the elevator to return to frame 0
before trying to walk onto it. If someone else has taken the elevator up to
frame 1 and the bot happens to approach the elevator before it returns to frame
0, the bot will fall into the shaft below the elevator rather than waiting for
the elevator to return. The bot will collect the powerup and then become 'lost'
after not reaching the next node in time. Then, the bot will start taking the
recovery path that starts under the elevator. The bot will activate the
elevator switch to send the elevator up to frame 1, then the bot will jump out
of the elevator shaft and return to the normal path for boarding the elevator.
To mitigate an infinite loop of jumping out of the shaft and falling back into
it, the rbot-arrival.cog has been used to make the bot wait at the end of the
recovery path for the elevator to return to frame 0. Then the bot will return
to the normal path and may ride the elevator up normally.