Refactor adjoin LOS check to improve performance on larger maps
For some of the larger maps, the recently-added LOS check for non-passable
adjoins was significantly hurting performance, especially as more bots were
added to the match. Investigation revealed that even replacing the LOS check
logic with a simple for-loop that counted from 0 to 5000 and did nothing else
still had a significant performance impact on the game (that's 5000 loop
iterations per powerup/target per AI pulse per bot). With the maximum number
of bots enabled, the game was unplayable due to horrible framerate.

To mitigate this performance issue, the adjoin LOS check has been
refactored to scan only the adjoins in the sectors that are immediately
adjacent to the bot and the pwoerup/target, rather than checking every
adjoin in the map every time. This revised implementation will detect
blocked LOS up to two adjoins away from either the bot or the other thing.
Although this is unable to detect blocked LOS for adjoins that are farther
away than two hops, this should be sufficient to handle the most common
cases. If necessary, a customizer cog can be used to handle any special
cases that are missed by this revised logic.

The revised implementation was tested in Blades of Death, Spaceport, and The
Challenge at Nar Shaddaa with the maximum number of bots (16 for JK, 32 for
MotS). No performance hits were observed with the final revision to the code.
However, to increase the stress test intensity, the code was temporarily
tweaked to execute the LOS logic multiple times per call. Some minor framerate
reduction started to occur with the maximum number of bots enabled, but the
game was still playable. This may indicate that the revised implementation
performs well enough on a reasonably fast system but could possibly still
encounter performance issues in some extreme cases or on slower systems.