vaporator: Improve navigation to supershield behind crates
Bots now promptly jump out from behind the crates after collecting the
supershield.

Node placement and jump delay timing has been tweaked for smoother navigation.

The path to the supershield now spurs off from the main path instead of the
switch door path. Bots should not need to be using that special path in order
to collect the supershield.

The weight of the supershield path has been reduced slightly so bots only
occasionally choose to look for the supershield.