Work around maximum move speed limitation to fix force speed
Apparently, the maximum movespeed value accepted by AiSetMoveSpeed is 2.0. To
work around this, the maxthrust value in the bot's template has been increased.
The maxthrust and movespeed values are multiplied together by the game engine
when computing acceleration for an actor's movement, so having a larger
maxthrust value allows smaller movespeed values to be passed in to
AiSetMoveSpeed to achieve the same movement speed while avoiding the maximum
movespeed value limit enforced by that verb.