throne: Add some jumps on the catwalk paths for smoother bot navigation
Bots now proactively jump from the catwalks to the hallways instead of relying
on the obstruction jump. The 0.1 second delay allows the nodes to be slightly
farther out on the catwalk (rather than stright down from the hallway edge) and
also gives bots that respawn on the catwalk a moment to face the correct
direction before attempting to jump up into the hallway.