Refactor rbot-srv.cog to await signal from arrival cog instead of event ID
This change eliminates the need for arrival cogs to adhere to some arbitrary
scheme for generating distinct event IDs to mitigate collisions with event IDs
generated by different arrival cogs. Instead, the cogs provide an arrival ID,
which need only be distinct for that cog instance. When the arrival trigger is
received by srv, it checks to see if the expected arrival cog sent it, which
prevents collisions with arrival IDs from other cogs.