vaporator: Fix activation sound distance when bot activates door switch
The mdm07_doorswitch.cog is a modified clone of 00_doorswitch.cog that
increases the on/off sound distance. While updating this cog to support Rbots,
the activate trigger handling code was copied from rbot-door-switch.cog but
apparently was not adjusted to match the modified activate code already in the
cog. That is, when bots would activate the door switch, the sound would play
with default distance rather than the increased distance. This discrepancy is
now fixed.