Remove unnecessary distance check from rbot-activate.cog
This distance check is meant to ensure that the bot cannot erroneously activate
something if they are not close enough to it. This was vital in earlier
revisions of the activate trigger code, as every cog needed to determine if the
bot was trying to activate one of its switches and also determine which switch
was being activated.

In the current implementation, the activation event already knows the specific
switch that the bot needs to activate, Since the event only fires upon the bot
reaching a specific node (which should always be close to the switch), the
distance check is no longer really necessary. In fact, if a cog is configured
incorrectly, the check could even cause inconsistent behavior, potentially
making the problem harder to troubleshoot. The rbot-arrival.cog omits the
distance check for this reason, in order to mitigate bots waiting around after
failing to activate the switch.

With these considerations in mind, the distance check has been removed.