Add optional CTF/KFY activation parameters to rbot-activate.cog
The new parameters in the CTF/KFY variants of rbot-activate.cog enable
conditional activation of the switch based on whether the CTF flags or
Ysalamiri have been taken or not. These parameters are optional.

For simplicity, giving the parameters a value of 1 will match both the
CTF/KFY capture states of 1 (dropped) and 2 (taken/held). There will
usually be no need to distinguish between these two cases, and in cases
where it would be useful, a custom cog may be required anyway. This
avoids the need for a map to configure two instances of the cog with
otherwise identical parameters just to match both states.