Fix some minor bot arrival issues
While a bot is awaiting a pending arrival event, progress toward the
next node will no longer be checked. That is, the bot will not reach
the next node until after the arrival event has occurred (or is
cleared), even if the bot is close enough to have reached the node.
This prevents the "bot lost" timer from being reset by reaching the
node before the arrival event fires (the pending arrival overrides the
timer duration with the arrival timeout). For example, this could occur
in the Bespin Mining Station map when the bot uses the reactor thrust
trap due to the path nodes being extremely close to each other.

Using the Rbot Admin function to teleport a bot now properly clears any
pending arrival event state.