Enhance rbot-arrival.cog to keep retrying activation
While a bot is waiting for arrival of the thing, he will periodically check to
see if the thing is heading to the desired frame. If it is not, the bot will
try to activate the button/console again to call the thing to the desired
frame. This check will occur periodically until either the thing arrives or the
timeout elapses.

The rbot-arrival.cog now supports the buttoncel parameter to help minimize
unnecessary activation triggers (e.g. if the thing is not moving to the frame
but the switch is also not in a state to be activated).

This enhancement is intended to help bots recover in situations where they
arrive at a node and need to activate a switch to call a thing to another
frame, but the switch is not yet in a state where the bot is able to activate
it. For example, when the bot needs to call an elevator but the elevator is
currently moving to another floor, the retry timer will eventually detect that
the elevator can be called back and the bot will activate the switch then.
Without the retry timer, the bot would simply wait at the node until the
arrival timeout elapses and then try to call the elevator again after the nav
system sends the "lost" bot back to the same node (assuming the bot's run
direction is not reversed by the nav update).