throne: Improve pit detection to use world geometry and vector math
The previous pit detection solution relied on manually annotating sectors with
arbitrary zone numbers and preventing the bots from crossing between zones to
seek items and targets. This worked well enough and the logic was simple, but
it placed more burden on the map maintainer to manually specify the many
sectors that make up all the zones.

This revised solution instead relies on the map maintainer to simply identify
the pits themselves. This is accomplished by specifying the adjoins through
which the bots could fall into each pit, which is arguably a more intuitive and
user-friendly way to annotate the hazard. The customizer cog inspects the world
geometry around each pit and uses vector math to determine if a bot would cross
over a pit to reach an item or target. As with a zone mismatch, the bot will be
prevented from moving toward an item or target if there is a pit in the way.